Add a `setup.py` file
Add a `setup.py` file and reorganize the code into a `vtctrans` module to allow installation via `pip`.
